+++ ios/App/App/AvantaBridgeViewController.swift
@@ -1,0 +1,9 @@
+import Capacitor
+
+@objc(AvantaBridgeViewController)
+class AvantaBridgeViewController: CAPBridgeViewController {
+    override func capacitorDidLoad() {
+        bridge?.registerPluginType(SecureOAuthPlugin.self)
+    }
+}
+